+++ frontend/reducers/index.js
@@ -1,11 +1,16 @@
 import { UPDATE_CHECKOUT_IS_ORDERABLE } from '../constants';
+
+const initialState = {
+  isOrderable: true,
+};
+
 /**
  * State to be used by checkout button for terms and conditions
  * @param {Object} state state
  * @param {Object} action action
  * @returns {Object}
  */
-const isOrderableByTerms = (state = { isOrderable: true }, action) => {
+const isOrderableByTerms = (state = initialState, action) => {
   switch (action.type) {
     case UPDATE_CHECKOUT_IS_ORDERABLE:
       return {
@@ -13,9 +18,7 @@ const isOrderableByTerms = (state = { isOrderable: true }, action) => {
         isOrderable: action.isOrderable,
       };
     default:
-      return {
-        ...state,
-      };
+      return state;
   }
 };