+++ CHANGELOG.md
@@ -2,10 +2,63 @@
  All notable changes to this project will be documented in this file.
  The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### [1.1.0] - 2019-03-07
+- Added support for specific product Id configurations and variant products based on base product Id. Example config:
+```json
+{
+  "checkboxValues": {
+    "type": "admin",
+    "destination": "frontend",
+    "default": [
+      {
+        "displayOn": "all products",
+        "productCheckboxValues": [
+          {
+            "text": "Required checkbox for all products",
+            "textColor": "#000"
+          },
+          {
+            "text": "Additional required checkbox for all products",
+            "textColor": "#000"
+          }
+        ]
+      },
+      {
+        "displayOn": "001",
+        "productCheckboxValues": [
+          {
+            "text": "Required checkbox for 001",
+            "textColor": "#000"
+          },
+          {
+            "text": "Additional required checkbox for 001",
+            "textColor": "#000"
+          }
+        ]
+      },
+      {
+        "displayOn": "002",
+        "productCheckboxValues": [
+          {
+            "text": "Required checkbox for 002",
+            "textColor": "#000"
+          }
+        ]
+      }
+    ],
+    "params": {
+      "required": false,
+      "type": "json",
+      "label": "Terms and conditions to apply for specific product Id's or all products"
+    }
+  }
+}
+```
+
 ### [1.0.1] - 2019-02-14
 Updated peerDependencies.
 
- ### [1.0.0] - 2019-02-06
+### [1.0.0] - 2019-02-06
 First version of the extension.
 ### Added
 - Added Terms and Conditions Checkboxes that must be checked in order to checkout.